+++ ev.c
@@ -1405,8 +1405,8 @@ const char* ev_strerror(int err)
 #line 23 "ev.c"
 ////////////////////////////////////////////////////////////////////////////////
 // FILE:    ev/fs.c
-// SIZE:    23436
-// SHA-256: 5d781978fa7050ea07c0bb3aa9e0b476a96a91eb902cf117d4b9b822117fde47
+// SIZE:    23492
+// SHA-256: 052abe1cfd800cadeb7ea72bc2ff898b4be983fdc4ca0656184d60772f663c4a
 ////////////////////////////////////////////////////////////////////////////////
 #line 1 "ev/fs.c"
 #include <sys/stat.h>
@@ -1803,7 +1803,7 @@ static void _ev_file_on_write(ev_work_t* work)
     ev_file_t* file = req->file;
     ev_write_t* write_req = &req->req.as_write.write_req;
 
-    req->result = ev_file_write_sync(file, write_req->bufs,
+    req->result = ev__fs_writev(file->file, write_req->bufs,
         write_req->nbuf);
 }
 
@@ -2130,15 +2130,16 @@ ssize_t ev_file_pread(ev_file_t* file, ev_fs_req_t* req, ev_buf_t bufs[],
     return _ev_file_read_template(file, req, bufs, nbuf, offset, cb, _ev_file_on_pread);
 }
 
-int ev_file_write(ev_file_t* file, ev_fs_req_t* req, ev_buf_t bufs[],
+ssize_t ev_file_write(ev_file_t* file, ev_fs_req_t* req, ev_buf_t bufs[],
     size_t nbuf, ev_file_cb cb)
 {
-    return _ev_file_pwrite_template(file, req, bufs, nbuf, 0, cb, _ev_file_on_write);
-}
+    if (file->base.loop == NULL)
+    {
+        EV_ASSERT(req == NULL && cb == NULL, "file open in synchronous mode.");
+        return ev__fs_writev(file->file, bufs, nbuf);
+    }
 
-ssize_t ev_file_write_sync(ev_file_t* file, ev_buf_t bufs[], size_t nbuf)
-{
-    return ev__fs_writev(file->file, bufs, nbuf);
+    return _ev_file_pwrite_template(file, req, bufs, nbuf, 0, cb, _ev_file_on_write);
 }
 
 int ev_file_pwrite(ev_file_t* file, ev_fs_req_t* req, ev_buf_t bufs[],