+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 2. rename `ev_file_exit()` to `ev_file_close()`
 3. merge `ev_file_read_sync()` with `ev_file_read()`
 4. merge `ev_file_pread_sync()` with `ev_file_pread()`
+5. merge `ev_file_write_sync()` with `ev_file_write()`
 
 ### Bug Fixes
 1. `ev_hrtime()` no longer require initialize event loop first.